Create switchvox-usb.pl
Short and somewhat ugly script to take a Switchvox ISO file, and prepare it for booting from a usb drive. Works to provide a stand-alone .img file intended for a 'dd' to the drive, or to create an archive file for integration in an already prepped syslinux bootable thumb drive.