updating fsck-all, fixing bugs
add short and trim options to fsck-all.
short removes 'cck'/the bad block check from the procedure to save time
trim adds the '-E discard' TRIM operation to the check, which is really
only useful on SSDs that are not under a RAID configuration, as most
RAID cards don't support TRIM

fixing a bug in smart testing which leaves in unintended spaces, causing
the serial number lookup to fail for some drives.